training queue recommendation, completion

+++ src/routes/dashboard/vaccs/[id]/training/queues/manage/data-table-actions.svelte
@@ -3,6 +3,7 @@
   import { Button } from "$lib/components/ui/button";
   import { Copy, MoreHorizontal } from "lucide-svelte";
   import { page } from "$app/stores";
+  import {can} from "$lib/perms/can";
 
   export let id: string;
 </script>
@@ -24,16 +25,36 @@
       <DropdownMenu.Item on:click={() => navigator.clipboard.writeText(id)}>
         Copy ID
       </DropdownMenu.Item>
-      <DropdownMenu.Item
-        href="/dashboard/vaccs/{$page.params.id}/training/queues/{id}/edit/">
-        Edit details
-      </DropdownMenu.Item>
     </DropdownMenu.Group>
     <DropdownMenu.Separator />
-    <DropdownMenu.Item
-      href="/dashboard/vaccs/{$page.params.id}/training/queues/{id}/delete/"
-      class="text-red-500 data-[highlighted]:bg-red-500/30 data-[highlighted]:text-red-500">
-      Delete queue
-    </DropdownMenu.Item>
+
+    {#if can($page.data.roles, $page.params.id, $page.data.user.vaccId, `vacc.${$page.params.id}.training.queues.recommend`)}
+      <DropdownMenu.Group>
+        <DropdownMenu.Item href="/dashboard/vaccs/{$page.params.id}/training/queues/{id}/recommend">
+          Recommend Student
+        </DropdownMenu.Item>
+        <DropdownMenu.Item href="/dashboard/vaccs/{$page.params.id}/training/queues/{id}/complete">
+          Mark Student Complete
+        </DropdownMenu.Item>
+      </DropdownMenu.Group>
+
+      <DropdownMenu.Separator />
+    {/if}
+
+    {#if can($page.data.roles, $page.params.id, $page.data.user.vaccId, `vacc.${$page.params.id}.training.queues.manage`)}
+      <DropdownMenu.Group>
+        <DropdownMenu.Item
+                href="/dashboard/vaccs/{$page.params.id}/training/queues/{id}/edit/">
+          Edit details
+        </DropdownMenu.Item>
+        <DropdownMenu.Item
+                href="/dashboard/vaccs/{$page.params.id}/training/queues/{id}/delete/"
+                class="text-red-500 data-[highlighted]:bg-red-500/30 data-[highlighted]:text-red-500">
+          Delete queue
+        </DropdownMenu.Item>
+      </DropdownMenu.Group>
+    {/if}
+
+
   </DropdownMenu.Content>
 </DropdownMenu.Root>
